CBL-6456 : Implement Swift LogSink API (#3358)
* CBL-6456 : Implement Swift LogSink API

* Implemented Swift LogSink API.

* Fixed using out-of-scope CBLStringBytes when setting directory and header to C4 API for CBLFileLogSink ; Added an option to CBLStringBytes to disable inline storage.

* Ran the script to generate public symbol for the logging API.

* Renamed domain property to domains in CBLConsoleLogSink and CBLCustomLogSink.

* Changed maxKepthFiles to NSInteger and maxFileSize to long long in CBLFileLogSink.

* Updated and completed missing API docs in Objective-C LogSink API.

* Used @testable import so that the tests can use the internal function.

+++ Objective-C/CBLDefaults.m
@@ -42,7 +42,7 @@ const NSInteger kCBLDefaultLogFileMaxRotateCount = 1;
 
 const BOOL kCBLDefaultFileLogSinkUsePlaintext = NO;
 
-const uint64_t kCBLDefaultFileLogSinkMaxSize = 524288;
+const long long kCBLDefaultFileLogSinkMaxSize = 524288;
 
 const NSInteger kCBLDefaultFileLogSinkMaxKeptFiles = 2;
